tiding up with structs file
